creating component for filters
