+++ src/app/views/query/index.js
@@ -13,6 +13,7 @@ import fimsMap from '../../components/map';
 import fimsQuery from './query.component';
 import fimsQueryForm from './form/query-form.component';
 import fimsQueryTable from './table/query-table.component';
+import fimsFilterButtons from './form/filter-buttons/filter-buttons.component';
 
 const dependencies = ['sly', router, QueryService, projectService, fimsMap];
 
@@ -21,4 +22,5 @@ export default angular
   .run(routing) // need to declare in run block. otherwise $transitions is not available
   .component('fimsQuery', fimsQuery)
   .component('fimsQueryForm', fimsQueryForm)
-  .component('fimsQueryTable', fimsQueryTable).name;
+  .component('fimsQueryTable', fimsQueryTable)
+  .component('fimsFilterButtons', fimsFilterButtons).name;
